// src/models/account.ts
export type Account = {
    _id: string,
    code: string;
    firmName: string;
    proprietor: string;
    phone: string;
    dno: string;
    strtNo: string;
    area: string;
    town: string;
    dl1: string;
    dl2: string;
    gst: string;
    mailid: string;
    opngBalAmt: string|number;
    groupCode: string|undefined;
}